Navigation Bar, Learn More button, Event section jump to specific section, cards in Home page with information and links
All navigation bars are in the same order, plus instead of having a home button I made the logo to lead to the home page.

Learn More on the Youth Club is fixed and jumps to the section with the buttons so people can read more about the program and tryouts. Direct jump from the home page is now an option too.

Reveal cards now have text and links to the other pages.

Minor clean up in the css for the navigation bars for Contact Us, Adult Club, and Gallery.

I have used the footer for the Gallery Page too.